LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowobjectfrom.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
  <body>
    <script src="../resources/securitypolicy-tests-base.js"></script>
    <script>
      if (document.SecurityPolicy.allowsObjectFrom('http://example.com/'))
          log('PASS object is allowed when no policy exists.');
      else
          log('FAIL object is not allowed when no policy exists.');

      injectPolicy("object-src http://notexample.com;");

      if (!document.SecurityPolicy.allowsObjectFrom('http://example.com/'))
          log('PASS object is not allowed when policy exists.');
      else
          log('FAIL object is allowed when policy exists.');
    </script>
  </body>
</html>
