LayoutTests/http/tests/security/contentSecurityPolicy/1.1/securitypolicy-allowobjectfrom-expected.txt

PASS object is allowed when no policy exists.
PASS object is not allowed when policy exists.
